RGM foi alterado para tipo Long

+++ src/main/java/laboratorioPedag/UMC_TCC_BACKEND/usuario/model/Usuario.java
@@ -31,6 +31,6 @@ public class Usuario implements Serializable {
     @NonNull
     private String senha;
     @NonNull
-    private double rgm;
+    private Long rgm;
 
 }
